refactor: Move TestGitServer into pkg/git
